[#2696] hammer.py: check Popen.poll() at the end
There was a wrong assumption that the first call to Popen.poll()
is always None. This is not the case when the process is fast enough
to finish before it gets polled, in which case it is not None, but
it is the return code instead. This caused some executes to have
empty output.